Move useFileDrop hook
